refactor(global): removed code smells
