+++ src/infra/repository/squad-configuration-repository.ts
