+++ src/messaging/handler.ts
@@ -4,7 +4,7 @@ import {
 } from 'aws-lambda';
 import { Logger } from 'tslog';
 import ConsumeSquadReportProcessMessagesUseCase from '../use-case/consume-squad-report-process-messages-use-case';
-import ProduceSquadReportProcessMessagesUseCase from '../use-case/produce-squad-report-process-messages-use-case';
+import ProduceSquadReportProcessMessagesUseCase from '../use-case/produce-report-process-messages-use-case';
 
 const produceSquadReportProcessMessagesUseCase = new ProduceSquadReportProcessMessagesUseCase();
 const consumeSquadReportProcessMessagesUseCase = new ConsumeSquadReportProcessMessagesUseCase();
